Move v050 gen xml test, re-export top lev
