+++ src/dep/v0_4_0/mod.rs
@@ -19,21 +19,21 @@ pub mod scpcoa;
 pub mod timeline;
 
 use crate::to_usize;
-use antenna::Antenna;
-use collection_info::CollectionInfo;
-use error_statistics::ErrorStatistics;
-use geo_data::GeoData;
-use grid::Grid;
-use image_creation::ImageCreation;
-use image_data::ImageData;
-use image_formation::{ImageFormation, RgAzComp, Rma};
-use match_info::MatchInfo;
-use pfa::Pfa;
-use position::Position;
-use radar_collection::RadarCollection;
-use radiometric::Radiometric;
-use scpcoa::ScpCoa;
-use timeline::Timeline;
+pub use antenna::Antenna;
+pub use collection_info::CollectionInfo;
+pub use error_statistics::ErrorStatistics;
+pub use geo_data::GeoData;
+pub use grid::Grid;
+pub use image_creation::ImageCreation;
+pub use image_data::ImageData;
+pub use image_formation::{ImageFormation, RgAzComp, Rma};
+pub use match_info::MatchInfo;
+pub use pfa::Pfa;
+pub use position::Position;
+pub use radar_collection::RadarCollection;
+pub use radiometric::Radiometric;
+pub use scpcoa::ScpCoa;
+pub use timeline::Timeline;
 
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct SicdMeta {
