+++ src/v1_3_0/mod.rs
@@ -22,15 +22,14 @@ pub use crate::dep::v0_4_0::{
     IdxLL, IdxLLH, IdxRowCol, IdxXyzPoly, Poly1D, Poly2D, RowCol, XyzPoly, CMPLX, LL, LLH, XYZ,
 };
 pub use crate::dep::v0_5_0::grid::Grid;
-
-use antenna::Antenna;
-use error_statistics::ErrorStatistics;
-use geo_data::GeoData;
-use image_formation::{ImageFormation, RgAzComp, Rma};
-use match_info::MatchInfo;
-use radar_collection::RadarCollection;
-use radiometric::Radiometric;
-use scpcoa::ScpCoa;
+pub use antenna::Antenna;
+pub use error_statistics::ErrorStatistics;
+pub use geo_data::GeoData;
+pub use image_formation::{ImageFormation, RgAzComp, Rma};
+pub use match_info::MatchInfo;
+pub use radar_collection::RadarCollection;
+pub use radiometric::Radiometric;
+pub use scpcoa::ScpCoa;
 
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct SicdMeta {
